i8042: Add debug_kbd option
A big problem with the current i8042 debugging option is that it outputs data
going to and from the keyboard by default. As a result, many dmesg logs
uploaded by users will unintentionally contain sensitive information such as
their password, as such it's probably a good idea not to output data coming
from the keyboard unless specifically enabled by the user.

Signed-off-by: Stephen Chandler Paul <cpaul@redhat.com>