psmouse: Export psmouse struct
We need to be able to access the psmouse struct from other drivers if we
want to be able to reroute the TrackPoint buttons on the T450 series to
the TrackPoint device, instead of just reporting them from the device
they're coming from (the touchpad). It's likely something else in the
future is going to need this anyway.

Signed-off-by: Stephen Chandler Paul <thatslyude@gmail.com>