tracing: Add support for enabling trace events in modules
For a while, passing events to enable via trace_enable on the kernel
cmdline has only actually worked for built-in modules. The reason for
this is that generally modules aren't loaded early enough to register
their tracing events, and we only enable events at two different spots
during the init process before we ever load any modules. This is rather
annoying for debugging things such as i915, where you might need to
enable register read/write tracing as the module is loading for an issue
that only occurs during the first modeset.

Signed-off-by: Lyude <cpaul@redhat.com>