drm/i915/skl: Move plane wms to state, remove redundant wms
Next step is to move all of the planes out of the very ugly struct we
have in drm_i915_private and into the individual plane states. This has
a lot of benefits:
 - No additional locking is needed to ensure watermarks are okay in
   atomic modesets; we can just rely on the modeset locks
 - No more playing the game of figuring out where to get skl_wm_values
   structs from.
 - Since we only keep the watermarks in plane states, they're always
   guaranteed to be up to date with the plane.

This also implies a couple of other changes:
 - Since the watermark values for planes are now guaranteed to be
   consistent, we don't have to do weird hacks with the plane's state to
   avoid writing blank watermarks. We just go ahead and write the plane
   watermark values unconditionally.

As well, we also no longer confusingly juggle between encoded watermark
values (for writing to the actual wm registers) and the actual pipe wm
values. We now *only* use skl_pipe_wm for storing watermarks, and use
the new skl_write_wm_level() helper for encoding and writing the
skl_pipe_wm structs to registers.

This removes a ton of code, and overall makes all of the skl wm code
easier to understand.

Signed-off-by: Lyude <cpaul@redhat.com>