drm/i915: Add intel_atomic_get_plane_state()
Since we're going to be moving the ddb into the plane states as well, we
need to make sure that we can always ensure that each plane has a state
we can store the ddb in during the atomic check phase. This is usually a
problem with the cursor, where we might leave it without a state during
the initial modeset.

Signed-off-by: Lyude <cpaul@redhat.com>