drm/i915: Skip retraining attempts if sink changes RX caps
Unfortunately this seems to be somewhat vaguely mentioned in the
DisplayPort spec, but there are multiple instances where a sink device
is advised to change it's maximum link rate and lane count in the DPCD.
An example of this is with downstream ports on non-MST aware sinks,
where version 1.1a of the spec says:

  "When a Branch Device detects a change in its downstream link through
   Hot Plug / Unplug detection (for example, a Sink Device gets plugged to
   one of the downstream ports), it must take the following actions within
   25 ms after the detection of the HPD pulse:

   - If the downstream link is DisplayPort, read the DPCD Receiver
     Capability field of the connected downstream DisplayPort Receiver and
     update the DPCD Receiver Capability field (MAX_LINK_RATE,
     MAX_LANE_COUNT, etc.) to comprehend the capability of the downstream
     DisplayPort Receiver. …"

This is hinted at during other parts of the spec, and additionally in
the MST spec is explicitly mentioned as valid behavior for MST hubs to
follow (we'll go more into this on a later patch).

So, be safe and save ourselves some time by skipping link retraining and
going straight to fallback link settings if we indicate that the RX caps
in the dpcd have changed from what they were before.

Signed-off-by: Lyude Paul <lyude@redhat.com>