drm/i915: Disable underruns during MST link retraining
Just like with SST retraining, Intel display hardware has a habit of
throwing underrun errors when a DP link requires retraining. The major
difference being that we can get underruns on any CRTCs that are driving
a sink in the given MST topology. This means that when we do MST
retraining, we have to go through and disable underrun detection on all
downstream connectors that have an active CRTC.

So, we introduce two simple helpers. intel_dp_get_active_crtc_mask() for
getting a mask of all active CRTCs on a given DP port, and
intel_dp_set_underrun_reporting() to enable/disable underrun detection
for all downstream sinks on a given DP port. For SST devices these
just operate on the attached drm_connector, for MST these act on every
downstream drm_connector in a topology that has a CRTC active on it.

Signed-off-by: Lyude Paul <lyude@redhat.com>