drm/i915: Implement proper fallback training for MST
For a while we actually haven't had any way of retraining MST links with
fallback link parameters like we do with SST. While uncommon, certain
setups such as my Caldigit TS3 + EVGA MST hub require this since
otherwise, they end up getting stuck in an infinite MST retraining loop.

MST retraining is somewhat different then SST retraining. For starters,
because there's never going to be a display running on the toplevel mst
parent the actual link retraining status isn't reported the same way as
SST. Every link retrain on the toplevel parent will report success, but
we can't actually know if the link retraining failed until we get an ESI
from the MST branch parent. This means that even if the hub knows we
need to fallback to a lower link rate, we'll never end up hitting the
fallback link parameters codepath in intel_dp_retrain().

So this patch adds a couple of necessary things to support this. For
starters, we start keeping track of how many times the hub has indicated
to us that the link status isn't healthy, and schedule
intel_dp_modeset_retry_work_fn() when we go above 5 retries.

Additionally, link training can fail in the middle of an atomic commit
that may not be entirely finished. For instance, if we're committing an
atomic state with two new displays on the MST port that is failing link
training, it's possible that link training could fail during enabling
the pipe for the first display and before we've started enabling the
pipe for the second display. To prevent racing with atomic commits in
this manner, we now also grab all of the appropriate drm_modeset locks
for each CRTC driving a downstream sink in
intel_dp_modeset_retry_work_fn() before changing connector status.

Signed-off-by: Lyude Paul <lyude@redhat.com>