drm/i915: Implement proper fallback training for MST
For a while we actually haven't had any way of retraining MST links with
fallback link parameters like we do with SST. While uncommon, certain
setups such as my Caldigit TS3 + EVGA MST hub require this since
otherwise, they end up getting stuck in an infinite MST retraining loop.

MST retraining is somewhat different then SST retraining. For starters,
because there's never going to be a display running on the toplevel mst
parent the actual link retraining status isn't reported the same way as
SST. Every link retrain on the toplevel parent will report success, but
we can't actually know if the link retraining failed until we get an ESI
from the MST branch parent. This means that even if the hub knows we
need to fallback to a lower link rate, we'll never end up hitting the
fallback link parameters codepath in intel_dp_retrain().

To make matters even more complicated: since every downstream CRTC
shares the same link, we can't actually do any kind of retraining until
we're in a state where it's possible for us to lock all downstream
CRTCs. But because of how intel's display hardware works, it's possible
for us to start getting false underruns from any of the display pipes
driving downstream sinks. So, in order to prevent the possibility of one
of the downstream CRTCs reporting underruns during the period where we
have to delay retraining until we can lock all of the downstream CRTCs,
we always disable underrun reporting on all of the CRTCs that are
available regardless of whether or not it's possible for us to perform
retraining just yet.

As well, since most of the intel DP short/long pulse handling has been
moved into dig work, we also rework MST retraining so that it happens in
a seperate worker so as to prevent our ESI handler from getting stalled
during retrains.

Additionally: we also take advantage of the new compute_deps hook so
that we can force userspace to include all downstream CRTCs whenever
userspace attempts to perform a modeset on a CRTC running on a DP MST
hub whose link requires fallback retraining.

Signed-off-by: Lyude Paul <lyude@redhat.com>