drm/i915: Make intel_dp_mst_atomic_check() idempotent
Part of MST retraining is going to require being able add additional
CRTCs to an atomic state after the initial atomic check computations
have been completed. Every time we add CRTCs, we need to be able to
rerun the atomic check sequence to make sure that the modeset is still
valid with the new CRTCs that have been added.

So, make intel_dp_mst_atomic_check() idempotent so that every time the
atomic check is rerun, any VCPI allocations we previously had assigned
to a CRTC in the state are freed before the recalculation to ensure that
we can avoid allocating virtual channels to a CRTC twice.

Signed-off-by: Lyude Paul <lyude@redhat.com>