drm/i915: Add intel_encoder->compute_deps() hook
When running multiple MST displays on the same MST topology, all of the
displays must be running at the same link rate always. If we were to
just simply call drm_dp_mst_topology_lower_link_rate() and allow
userspace to fix things without ensuring that multiple displays on the
same MST topology never run at different link rates, the following could
happen:

* Pipe A and B are running on the same MST topology. Mode is:
    A: active @ 5.4 GB/s
    B: active @ 5.4 GB/s
    C: inactive
* Channel retraining fails, MST topology must be retrained at a lower
  link rate. Set link status on respective DRM connectors to BAD
* Userspace notices that connectors have link status set to BAD,
  attempts the following modeset:
    A: active @ 5.4 GB/s -> active @ 2.7 GB/s
    B: active @ 5.4 GB/s
    C: inactive
* Pipe state mismatch (expected 2.7GB/s, got 5.4GB/s), etc. etc.

Here, the only acceptable way to perform this modeset that wouldn't
leave us in this scenario would be to pull B into the modeset and also
set it to 2.7GB/s, assuming that it's currently set mode along with A's
new mode still pass the atomic check.

Since doing this requires being able to pull new CRTCs into the atomic
state, let's add a simple hook to do this: compute_deps(). In the event
that an encoder detects that commiting a state would be impossible
without pulling in another CRTC into the state and redoing the atomic
check, the encoder can return false from compute_deps() which will force
i915 to reperform the atomic check with the new CRTCs added.

Signed-off-by: Lyude Paul <lyude@redhat.com>