drm/dp_mst: Add reset_state callback to topology mgr
This gives drivers subclassing drm_dp_mst_topology_state the ability to
prepare their topology states for a new hub to be connected whenever
it's detected that the currently connected topology has been
disconnected from the system. We'll need this in order to correctly
track RX capabilities in i915 from the topology's atomic state.

Signed-off-by: Lyude Paul <lyude@redhat.com>