drm/dp_mst: Add dp_mst_status debugfs node for all drivers
Originally I was just going to be adding dp_mst_status for nouveau until
Daniel Stone pointed out that we should probably just make this so it's
magically added for every DRM driver that's using the DRM DP MST
helpers. So, let's do that!

This makes it so we add a dp_mst_status debugfs node every time a
topology manager is initialized for a connector. Additionally, we

- Remove drm_dp_mst_dump_topology() and just move it's code into a
  private single_open() handler called
  drm_dp_mst_topology_mgr_debugfs_show()
- Remove i915's i915_dp_mst_status file, since it's not used by igt to
  my knowledge and can just be replaced with this

Signed-off-by: Lyude Paul <lyude@redhat.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>